[AOT Refactor] traced function transforms pt. 1
[ghstack-poisoned]
